Add acid fractionation

The temperature-dependent acid fractionation projection from 70 \(^\circ\)C to 25 \(^\circ\)C. It subtracts the aff from the D47-value. NOTE: The I-CDES temperature scale is at 70\(^\circ\)C by default, so for some labs acid fractionation correction is no longer needed.

acid_fractionation(
  .data,
  aff = 0.062,
  ...,
  D47 = D47_etf,
  D47_out = D47_final,
  quiet = default(quiet)
)

Arguments

.data

A tibble.

aff

Temperature-dependent acid fractionation projection from 70 \(^\circ\)C to 25 \(^\circ\)C. Defaults to 0.062, which is the average of the values obtained by De Vlieze et al., 2015 and Murray et al., 2016. See Muller et al., 2017.

...

These dots are for future extensions and must be empty.

D47

The column name of the \(\Delta_47\) values to use for the acid fractionation calculation.

D47_out

The desired new column name.

quiet

If TRUE, print info messages. If NULL (default) listens to options(clumpedr.quiet)

Value

Same as .data but with new column, named for the value of parameter D47_out.

References

W. F. Defliese, M.T. Hren, K. C. Lohmann. Compositional and temperature effects of phosphoric acid fractionation on \(\Delta_{47}\) analysis and implications for discrepant calibrations. Chem. Geol. 2016, 396, 51.

S. T. Murray, M. M. Arienzo, P. K. Swart. Determining the \(\Delta_{47}\) acid fractionation in dolomites. Geochem. Cosmochim. Acta 2016, 174, 42.

I. A. Muller, A. Fernandez, J. Radke, J. van Dijk, D. Bowen, J. Schwieters, S. M. Bernasconi. Carbonate clumped isotope analyses with the long-integration dual-inlet (LIDI) workflow: scratching at the lower sample weight boundaries. Rapid Commun. Mass Spectrom. 2017, 31, 1057–1066.

Examples

# generate artificial example data
dat <- tibble::tibble(D47 = rnorm(10))
dat <- dat |>
  acid_fractionation(aff = 0.062, D47 = D47, D47_out = D47_aff)
#> Info: adding temperature-dependent acid fractionation factor of 0.062.