Compute \(\delta^{13}\)C, \(\delta^{18}\)O \(\Delta_{47}\), \(\Delta_{48}\), and \(\Delta_{49}\) values

Compute \(\delta^{13}\)C, \(\delta^{18}\)O \(\Delta_{47}\), \(\Delta_{48}\), and \(\Delta_{49}\) values

bulk_and_clumping_deltas(
  .data,
  ...,
  d45 = d45,
  d46 = d46,
  d47 = d47,
  d48 = d48,
  d49 = d49,
  R13_PDB = default(R13_PDB),
  R18_PDB = default(R18_PDB),
  R17_PDBCO2 = default(R17_PDBCO2),
  R18_PDBCO2 = default(R18_PDBCO2),
  lambda = default(lambda),
  D17O = default(D17O),
  quiet = NULL
)

Arguments

.data

A tibble containing s44-s49, r44-r49, R45-R49, R45_wg-R49_wg, d45-d49, and working gas compositions d13C_PDB_wg and d18OPDBCO2_wg.

...

These dots are for future extensions and must be empty.

d45

Column name of d45.

d46

Column name of d46.

d47

Column name of d47.

d48

Column name of d48.

d49

Column name of d49.

R13_PDB

The R13 value of the PDB reference. Defaults to 0.01118, from Chang and Li, 1990.

R18_PDB

The R18 value of the PDB reference. Defaults to 1.008751, from Coplen et al., 1983.

R17_PDBCO2

The R17 value of PDB (g). Defaults to 0.0003931, from Assonov and Brenninkmeijer 2003, re-scaled by Brand, 2010.

R18_PDBCO2

The R18 value of PDB (g). Defaults to 0.00208839, derrived from Baertschi, 1976, calculated by Brand 2010 eq. 2.

lambda

Isotopic fractionation of \(^{17}\)O. Defaults to 0.528, from Barkan and Luz, 2005.

D17O

The difference between the expected \(\delta^{17}\)O (\(\delta^{18}\)O\(-\lambda*\delta^{18}\)O) and the actual \(\delta^{17}\)O value.

quiet

If TRUE, print info messages. If NULL (default) listens to options(clumpedr.quiet)

Value

Same as .data but with new columns R45-R49, a lot of intermediate steps, and D47_raw.

References

Daeron, M., Blamart, D., Peral, M., & Affek, H. P., Absolute isotopic abundance ratios and the accuracy of \(\Delta_{47}\) measurements, Chemical Geology 2016, 442, 83–96. doi:10.1016/j.chemgeo.2016.08.014

Examples

# generate some artificial data to demonstrate:
x <- function() rnorm(n = 10)
dat <- tibble::tibble(
 s44 = x(), s45 = x(), s46 = x(), s47 = x(), s48 = x(), s49 = x(),
 r44 = x(), r45 = x(), r46 = x(), r47 = x(), r48 = x(), r49 = x(),
 d45 = x(), d46 = x(), d47 = x(), d48 = x(), d49 = x(),
 d13C_PDB_wg = x(), d18O_PDBCO2_wg = x())
dat <- dat |>
  bulk_and_clumping_deltas()
#> Info: calculating δ¹³C, δ¹⁸O, and Δ's.