Calculate the Empirical Transfer Function

Calculate the Empirical Transfer Function

calculate_etf(
  .data,
  ...,
  raw = D47_raw_mean,
  exp = expected_D47,
  session = Preparation,
  etf = etf,
  etf_coefs = etf_coefs,
  slope = slope,
  intercept = intercept,
  parallel = FALSE,
  quiet = NULL
)

Arguments

.data

A tibble.

...

These dots are for future extensions and must be empty.

raw

Column name of raw \(\Delta_{47}\) values.

exp

Column name of expected \(\Delta_{47}\) values.

session

The column name to group analyses by. Defaults to Preparation.

etf

The column name of the new model.

etf_coefs

The column name with the coefficients of the model.

slope

The column name of the new slope.

intercept

The column name of the new intercept.

parallel

Whether or not (default) to process this in parallel, using package furrr.

quiet

If TRUE, print info messages. If NULL (default) listens to options(clumpedr.quiet)