Clean up the did file info

Converts mass spectrometer metadata into the correct types and appends the initial intensity of mass 44.

clean_did_info(
  .did,
  masspec = NULL,
  ...,
  std_names = c(paste0("ETH-", 1:4), "IAEA-C2", "Merck"),
  oth_name = "other",
  quiet = NULL
)

Arguments

.did

An iso file, resulting from isoreader::iso_read_dual_inlet().

masspec

The name of the mass spectrometer to append.

...

These dots are for future extensions and must be empty.

std_names

Character vector of the standard names to find in Identifier 1.

oth_name

Single general name to assign to non-standard "other" measurements.

quiet

If TRUE, print info messages. If NULL (default) listens to options(clumpedr.quiet)

See also