Default parameters

Default parameters

Arguments

R13_PDB

The R13 value of the PDB reference. Defaults to 0.01118, from Chang and Li, 1990.

R18_PDB

The R18 value of the PDB reference. Defaults to 1.008751, from Coplen et al., 1983.

R17_PDBCO2

The R17 value of PDB (g). Defaults to 0.0003931, from Assonov and Brenninkmeijer 2003, re-scaled by Brand, 2010.

R18_PDBCO2

The R18 value of PDB (g). Defaults to 0.00208839, derrived from Baertschi, 1976, calculated by Brand 2010 eq. 2.

lambda

Isotopic fractionation of \(^{17}\)O. Defaults to 0.528, from Barkan and Luz, 2005.

D17O

The difference between the expected \(\delta^{17}\)O (\(\delta^{18}\)O\(-\lambda*\delta^{18}\)O) and the actual \(\delta^{17}\)O value.

D47

TODO: look up

D48

TODO: look up

D49

TODO: look up

References

Assonov, S., Brenninkmeijer C. On the \(^{17}\)O correction for CO\(_{2}\) mass spectrometric isotopic analysis Rapid Commun. Mass Spectrom., 2003, 17, pp. 1007-1016, 10.1002/rcm.1012

Baertschi, P., Absolute \(^{18}\)O content of standard mean ocean water Earth Planet. Sci. Lett., 1976, 31, pp. 341-344

Barkan, E., Luz B. High-precision measurements of \(^{17}\)O/\(^{16}\)O and \(^{18}\)O/\(^{16}\)O ratios in H2O Rapid Commun. Mass Spectrom., 2005, 19, pp. 3737-3742

Chang, T., Li W. A calibrated measurement of the atomic weight of carbon Chin. Sci. Bull., 1990, 35 (290)

Coplen, T., C. Kendall, Hopple J. Comparison of stable isotope reference samples Nature, 1983, 302 (5905) , pp. 236-238, 10.1038/302236a0

Brand, W. A., Assonov, S. S., & Coplen, T. B. Correction for the \(^{17}\)O interference in \(\delta^{13}\)C measurements when analyzing CO\(_{2}\) with stable isotope mass spectrometry (IUPAC Technical Report). Pure and Applied Chemistry, 2010, 82(8), 1719--1733. doi:10.1351/PAC-REP-09-01-05