Find outliers

This function finds outliers based on several criteria. Here, we define an outlier as a measurement that has:

an initial mass 44 intensity below init_low.

an initial mass 44 intensity above init_high.

an imbalance between sample and reference gas mass 44 intensities of more than diff.

a clumped value that is more than nsd_off standard deviations away from the mean.

find_outliers(
  .data,
  ...,
  init_low = 8000,
  init_high = 40000,
  init_diff = 1200,
  param49_off = 1,
  internal_sd = 0.15,
  n_min = 5,
  n_id1 = 5,
  nsd_off = 4,
  D47 = D47_raw,
  session = Preparation,
  id1 = `Identifier 1`,
  quiet = NULL
)

Arguments

.data

A tibble with raw Delta values and file information.

...

These dots are for future extensions and must be empty.

init_low

Minimum initial intensity threshold for mass 44.

init_high

Maximum initial intensity threshold for mass 44.

init_diff

Maximum initial difference in mass 44 threshold between standard and sample gas.

param49_off

Where param_49 > this value is considered an outlier.

internal_sd

The internal standard deviation outlier cutoff.

n_min

Minimum number of aliquots within session to calculate threshold.

n_id1

Minimum number of aliquots within session to calculate threshold within group.

nsd_off

Number of standard deviations away from the mean threshold.

D47

The column with \(\Delta_{47}\) values.

session

Column name that defines correction session.

id1

Column name of the sample/standard identifier.

quiet

If TRUE, print info messages. If NULL (default) listens to options(clumpedr.quiet)