Find session outliers.

Find outliers that are more than 4 standard deviations away from the session median.

find_session_outlier(
  .data,
  ...,
  n = 5,
  nsd_off = 4,
  D47 = D47_raw,
  outlier_session = outlier_session_D47,
  session = Preparation,
  quiet = NULL
)

Arguments

.data

A tibble with raw Delta values and file information.

...

These dots are for future extensions and must be empty.

n

The minimum number of measurements in the session needed to calculate an offset from the median.

nsd_off

The number of standard deviations away from the median.

D47

The column to calculate the internal sd value for.

outlier_session

Column name of new output column.

session

The session for which to calculate the standard deviation and median values.

quiet

If TRUE, print info messages. If NULL (default) listens to options(clumpedr.quiet)