Compute isotopic ratios

Compute isotopic ratios for a sample with isotopic ratios R13 and R18, optionally accounting for non-zero values of \(\Delta^{17}\)O and clumped isotope anomalies (all expressed in permil). Function based on Daeron et al., 2016.

isobar_ratios(
  .data,
  ...,
  R13 = R13,
  R18 = R18,
  R45 = R45,
  R46 = R46,
  R47 = R47,
  R48 = R48,
  R49 = R49,
  R17_PDBCO2 = default(R17_PDBCO2),
  R18_PDBCO2 = default(R18_PDBCO2),
  lambda = default(lambda),
  D17O = default(D17O),
  D47 = default(D47),
  D48 = default(D48),
  D49 = default(D49)
)

Arguments

.data

A tibble with columns R13 and R18.

...

These dots are for future extensions and must be empty.

R13

The column name for R13.

R18

The column name for R18.

R45

The output column name for R45.

R46

The output column name for R46.

R47

The output column name for R47.

R48

The output column name for R48.

R49

The output column name for R49.

R17_PDBCO2

The R17 value of PDB (g). Defaults to 0.0003931, from Assonov and Brenninkmeijer 2003, re-scaled by Brand, 2010.

R18_PDBCO2

The R18 value of PDB (g). Defaults to 0.00208839, derrived from Baertschi, 1976, calculated by Brand 2010 eq. 2.

lambda

Isotopic fractionation of \(^{17}\)O. Defaults to 0.528, from Barkan and Luz, 2005.

D17O

The difference between the expected \(\delta^{17}\)O (\(\delta^{18}\)O\(-\lambda*\delta^{18}\)O) and the actual \(\delta^{17}\)O value.

D47

TODO: look up

D48

TODO: look up

D49

TODO: look up

References

Daeron, M., Blamart, D., Peral, M., & Affek, H. P., Absolute isotopic abundance ratios and the accuracy of \(\Delta_{47}\) measurements, Chemical Geology 2016, 442, 83--96. doi:10.1016/j.chemgeo.2016.08.014