Reverse temperature calibration

A clumped isotope temperature calibration in reversed form, where the calibration $$y = a \times 10^6 / T^2 + b$$ with \(T\) in Kelvin, is expressed as a function of \(T_C\) (in degrees Celsius): $$T_C = \sqrt{(a \times 10^6) / (y - b)} - 273.15$$

revcal(
  D47,
  slope = 0.0449,
  intercept = 0.167,
  slpcnf = 0.001,
  intcnf = 0.01,
  ignorecnf = FALSE
)

Arguments

D47

The \(\Delta_{47}\) value.

slope

The slope of the regression.

intercept

The intercept of the regression.

slpcnf

The confidence of the slope of the regression.

intcnf

The confidence of the intercept of the regression.

ignorecnf

Whether or not to ignore the confidence of the temperature calibration.

Value

A tibble with input \(\Delta_{47}\) and estimated Tc value with lower and upper bounds.

Details

Defaults to Bonifacie et al. 2017 $$\Delta_47 = (0.0449 \pm 0.001 \times 10^6) / T^2 + (0.167 \pm 0.01)$$

See also