Calculate the temperatures

This uses the revcal() function to calculate temperatures from \(\Delta_{47}\) values. At the moment it ignores the uncertainty in the regression.

temperature_calculation(
  .data,
  ...,
  D47 = D47_final,
  temp = temperature,
  slope = "slope",
  intercept = "intercept",
  quiet = NULL
)

Arguments

.data

A tibble, resulting from acid_fractionation().

...

These dots are for future extensions and must be empty.

D47

The column name of the \(\Delta_{47}\) values to use as input.

temp

The column name of the output temperature.

slope

Character(1) column name with the slope.

intercept

Character(1) column name with the intercept.

quiet

If TRUE, print info messages. If NULL (default) listens to options(clumpedr.quiet)

See also