Empirical Transfer Function

Calculate and apply the empirical transfer function (ETF). First append the expected values for the standards, then calculate the ETF per Preparation, then apply the ETF to all the samples and standards.

empirical_transfer_function(
  .data,
  ...,
  std_names = paste0("ETH-", 1:3),
  std_values = c(0.258, 0.256, 0.691) - 0.062,
  raw = D47_raw,
  exp = expected_D47,
  session = Preparation,
  id1 = `Identifier 1`,
  etf = etf,
  etf_coefs = etf_coefs,
  slope = slope,
  intercept = intercept,
  out = D47_etf,
  outlier = outlier,
  quiet = NULL,
  parallel = FALSE
)

Arguments

.data

A tibble resulting from collapse_cycles().

...

These dots are for future extensions and must be empty.

std_names

Names of the standards.

std_values

Expected values of the standards. Defaults to Bernasconi et al., 2021.

raw

Column name of raw \(\Delta_{47}\) values.

exp

Name of the new column that will hold expected values.

session

The column name to group analyses by. Defaults to Preparation.

id1

Column name of Identifier 1 (default).

etf

The column name of the new model.

etf_coefs

The column name with the coefficients of the model.

slope

The column name of the new slope.

intercept

The column name of the new intercept.

out

The new column name.

outlier

Column name of outlier (default).

quiet

If TRUE, print info messages. If NULL (default) listens to options(clumpedr.quiet)

parallel

Whether or not (default) to process this in parallel, using package furrr.

Details

This function is a wrapper for append_expected_values(), calculate_etf(), and apply_etf().

See also