nest cycle data

Nest the cycle data resulting from previous computations steps.

nest_cycle_data(
  .data,
  ...,
  masses = c(44:49, 54),
  ratios = c(paste0("s", masses), paste0("r", masses)),
  outlier_cycles = c(paste0("outlier_cycle_", c("low_s44", "low_r44", "high_s44",
    "high_r44", "drop_s44", "drop_r44", "s44", "r44")), "outlier_cycle"),
  outliers = "outlier",
  cycle_drop = paste0("cycle_", c("diff_s44", "diff_r44", "drop_s44", "drop_r44",
    "drop_num_s44", "drop_num_r44", "has_drop_s44", "has_drop_r44")),
  bgs = c(paste0("s44_bg", masses[!masses %in% c(44, 54)]), paste0("r44_bg",
    masses[!masses %in% c(44, 54)])),
  bg_corrected = paste0(ratios[!ratios %in% c("s44", "r44", "r54", "s54")], "_bg"),
  Rs = c("R45", "R46", "R47", "R48", "R49", "R45_wg", "R46_wg", "R47_wg", "R48_wg",
    "R49_wg", "R13_wg", "R18_wg", "R17", "R18", "R13"),
  deltas = c("d13C_PDB", "d18O_PDBCO2", "d18O_PDB", "d45", "d46", "d47", "d48", "d49"),
  isotopes = c("C12", "C13", "C16", "C17", "C18", "C626", "C627", "C628", "C636", "C637",
    "C638", "C727", "C728", "C737", "C738", "C828", "C838"),
  params = c("K", "A", "B", "C", "D", "aa", "bb", "cc"),
  stochastic = c("R45_stoch", "R46_stoch", "R47_stoch", "R48_stoch", "R49_stoch"),
  flags = c("R45_flag", "R46_flag"),
  Deltas = c("D45_raw", "D46_raw", "D47_raw", "D48_raw", "D49_raw"),
  p49 = "param_49",
  quiet = NULL
)

Arguments

.data

A tibble resulting from bulk_and_clumping_deltas().

...

Additional chacter vectors to nest by.

masses

The masses that are present in the mass spectrometer. Defaults to 44:49 and 54.

ratios

Ratio columns, based on masses.

outlier_cycles

Columns with outlier_cycle information.

outliers

Other outlier-related columns.

cycle_drop

Columns related to detecing cycle drops.

bgs

Backgrounds.

bg_corrected

Background-corrected raw intensities per cup.

Rs

R values.

deltas

delta values.

isotopes

C values.

params

Taylor polynomial parameters.

stochastic

Stochastic R values.

flags

Flag columns.

Deltas

Big delta values.

p49

Param 49.

quiet

If TRUE, print info messages. If NULL (default) listens to options(clumpedr.quiet)

Value

A summary of .data,

Details

The parameters are all strings with column names. They default to all the columns that have cycle information based on previous computation steps.